+++ adapters/template.pl
@@ -315,6 +315,13 @@ numbervars(Term, From, Next) :-
 	?????
 
 
+% '$lgt_directory_hash_as_atom'(+atom, -atom)
+%
+% returns the directory hash as an atom
+
+'$lgt_directory_hash_as_atom'(_, '').
+
+
 % '$lgt_compile_prolog_code'(+atom, +atom, +list)
 %
 % compile to disk a Prolog file, resulting from a