+++ core/core.pl
@@ -2516,7 +2516,7 @@ create_logtalk_flag(Flag, Value, Options) :-
 % versions, 'rcN' for release candidates (with N being a natural number),
 % and 'stable' for stable versions
 
-'$lgt_version_data'(logtalk(3, 2, 3, rc1)).
+'$lgt_version_data'(logtalk(3, 2, 3, rc2)).
 
 
 
@@ -5619,6 +5619,11 @@ create_logtalk_flag(Flag, Value, Options) :-
 	;	% assume absolute directory path
 		ObjectDirectory0 = ScratchDirectory
 	),
+	% append (if supported by the backend compiler) a directory hash value to the
+	% intermediate Prolog file name to try to avoid file name collisions when
+	% collecting all the intermediate files in the same directory for embedding
+	'$lgt_directory_hash_as_atom'(SourceDirectory, Hash),
+	atom_concat('_', Hash, UnderscoreHash),
 	% add a suffix based on the original extension to the file name to avoid
 	% intermediate and temporary file name conflicts when compiling two or
 	% more source files that share the same name but use different extensions
@@ -5627,7 +5632,8 @@ create_logtalk_flag(Flag, Value, Options) :-
 	;	sub_atom(SourceExtension, 1, _, 0, Suffix0) ->
 		atom_concat('_', Suffix0, Suffix)
 	),
-	atom_concat(SourceName, Suffix, ObjectName),
+	atom_concat(SourceName, UnderscoreHash, ObjectName0),
+	atom_concat(ObjectName0, Suffix, ObjectName),
 	% there must be a single object file extension defined in the Prolog adapter files
 	'$lgt_file_extension'(object, ObjectExtension),
 	atom_concat(ObjectName, ObjectExtension, ObjectBasename),